+++ tests/index.js
@@ -1,12 +1,2 @@
-// require('./routes/index'); // 6060
-// require('./middlewares/index'); // 6061
-// require('./services/index'); // 6062
-// require('./errorHandlers/index');
-// view engine
-
-require('./functionals/middlewares/index');
-require('./functionals/configurations/index');
-
-require('./units/config');
-require('./units/extracts');
-require('./units/generate');
+require('./functionals');
+// require('./units');